Refactor tests and add TODOs
- correct enum name
- fix sundry imports
- add missing c-tor
- add assertJ assertions
